Moved setUser (/api/setUser) method to the UserController
